Data Saves
since the buttons now work, we now have to start working on the main
game. The introduction is nearly done, with each line of text fading
in. Of course, I only have the intro for the forest right now but it
won’t be hard to add the other terrains. The big thing I did in this
commit is figure out have to save data, so that the user does not have
to restart the game all over again every time they close out the app
all the way and then reopen it. Next, I am going to try to finish out
all the intros, then perhaps figure out how to make this app run in
real time.