Finished intro animation and added inventory view
I have finished the animation for all the introductory text, at least
for the forest, but it won’t be hard to make it work for the other
terrains. Also added a collection view controller, which is how we will
display the inventory, but there is no code written for it yet.